:recycle: make event store implementation generic for all aggregates
