+++ marketplace-infrastructure/src/database/event_store/mod.rs
@@ -1,0 +1,5 @@
+mod event_store;
+pub use event_store::*;
+
+#[cfg(test)]
+mod test;
